fix: remove load event
Fixes KEYMANWEB-COM-7N.

This crash could happen if the page load event happened before KeymanWeb
had loaded and initialized.

The load event is no longer required as we have a minimum of version 14
of KeymanWeb, which has a Promise for post-initialization.

A secondary fix is to use `window.keyman` instead of plain `keyman`,
which guarantees a return of `undefined` instead of a `ReferenceError`.
There are some other calls that depend on this behaviour.